+++ kernel/include/defs.h
@@ -35,7 +35,7 @@ struct file*    filealloc(void);
 void            fileclose(struct file*, int drop_ofile_cnt);
 struct file*    filedup(struct file*);
 void            fileinit(void);
-int fileread(struct file *f, int user_dst, uint64 addr, size_t n, int64_t off);
+size_t fileread(struct file *f, int user_dst, uint64 addr, size_t n, int64_t off);
 int             filestat(struct file*, uint64 addr);
 int filewrite(struct file *f, int user_src, uint64 addr, size_t n, int64_t off);
